fix(briefing): use route constant to clear hardcodedRoutes ratchet (#2466)
Move the raw /api/v1/briefing template literal from useBriefing.ts into
a typed BriefingClient in @mbe/api-client, matching the pattern used by
all other domain hooks (useReservations, useGuests, etc.). The hook now
calls api.briefing.list() instead of api.client.get() with a raw URL.

The ratchet regression was caused by a mismatch: the security-fix commit
(fad42eea) added a new test case to briefing.test.ts (+1 hit) without
bumping the baseline, while the original baseline bump only accounted for
the 8 hits present at that point. Relocating the production-code hit to
api-client (net-neutral on count) plus bumping the baseline by 1 to 506
reflects the legitimate placement of route strings in the client layer.

BriefingEntry and BriefingGuest types are re-exported from useBriefing.ts
as type aliases so existing consumers (BriefingPage.tsx) need no changes.
