+++ apps/hospitality/src/hooks/useBriefing.ts
@@ -1,41 +1,11 @@
+import type { ApiBriefingEntry, ApiBriefingGuest } from "@mbe/api-client";
 import { createQueryHook } from "./create-query-hook.js";
 
 export const BRIEFING_QUERY_KEY = "briefing" as const;
 
-export interface BriefingGuest {
-  id: string;
-  name: string;
-  visitCount: number;
-  lastVisit: string | null;
-  dietaryRestrictions: string[] | null;
-  notes: string | null;
-  staffNotes: Array<{ text: string; createdBy: string; createdAt: string }>;
-  tags: string[] | null;
-  communicationPreference: string | null;
-}
-
-export interface BriefingEntry {
-  id: string;
-  date: string;
-  startTime: string;
-  endTime: string;
-  partySize: number;
-  status: "PENDING" | "CONFIRMED";
-  notes: string | null;
-  cancellationReason: string | null;
-  cancellationNote: string | null;
-  guestName: string | null;
-  guestId: string | null;
-  userId: string | null;
-  occasion: string | null;
-  seatingPreference: string | null;
-  tableId: string;
-  table?: { id: string; name: string; tableNumber: string | null } | null;
-  venueId: string | null;
-  createdAt: string;
-  updatedAt: string;
-  guest: BriefingGuest | null;
-}
+// Re-export api-client types under the names consumers already import
+export type BriefingGuest = ApiBriefingGuest;
+export type BriefingEntry = ApiBriefingEntry;
 
 export interface UseBriefingParams {
   date: string;
@@ -47,10 +17,7 @@ export const useBriefing = createQueryHook<BriefingEntry[], UseBriefingParams>({
   key: BRIEFING_QUERY_KEY,
   fetcher: async (params, api) => {
     if (!params) return [];
-    const response = await api.client.get<{ data: BriefingEntry[] }>(
-      `/api/v1/briefing?date=${encodeURIComponent(params.date)}&venueId=${encodeURIComponent(params.venueId)}`
-    );
-    return (response as { data: BriefingEntry[] }).data;
+    return api.briefing.list({ date: params.date, venueId: params.venueId });
   },
   getEnabled: (params) => Boolean(params?.date && params?.venueId),
 });
